Moved tests from controller to service.
